Splited classes into separate file
