+++ ReportPortal.ATM/APITests/Models/Body/AddWidgetBody.cs
@@ -1,4 +1,5 @@
-﻿using System.Text.Json.Serialization;
+﻿using System.Collections.Generic;
+using System.Text.Json.Serialization;
 
 namespace APITests.Models.Body;
 
@@ -16,32 +17,4 @@ public class AddWidgetBody
     public string Description { get; set; }
     [JsonPropertyName("filterIds")]
     public List<string> FilterIds { get; set; }
-}
-
-public class ContentParameters
-{
-    [JsonPropertyName("contentFields")]
-    public List<string> ContentFields { get; set; }
-    [JsonPropertyName("itemsCount")]
-    public string ItemsCount { get; set; }
-    [JsonPropertyName("widgetOptions")]
-    public WidgetOptions WidgetOptions { get; set; }
-}
-
-public class WidgetOptions
-{
-    [JsonPropertyName("zoom")]
-    public bool Zoom { get; set; }
-    [JsonPropertyName("timeLine")]
-    public string Timeline { get; set; }
-    [JsonPropertyName("viewMode")]
-    public string ViewMode { get; set; }
-}
-
-public class Filter
-{
-    [JsonPropertyName("value")]
-    public string Value { get; set; }
-    [JsonPropertyName("name")]
-    public string Name { get; set; }
 }
