+++ ReportPortal.ATM/APITests/Models/Body/ContentParameters.cs
@@ -1,0 +1,13 @@
+﻿using APITests.Models.Body;
+using System.Collections.Generic;
+using System.Text.Json.Serialization;
+
+public class ContentParameters
+{
+    [JsonPropertyName("contentFields")]
+    public List<string> ContentFields { get; set; }
+    [JsonPropertyName("itemsCount")]
+    public string ItemsCount { get; set; }
+    [JsonPropertyName("widgetOptions")]
+    public WidgetOptions WidgetOptions { get; set; }
+}
