+++ ReportPortal.ATM/APITests/Models/Body/Filter.cs
@@ -1,0 +1,9 @@
+﻿using System.Text.Json.Serialization;
+
+public class Filter
+{
+    [JsonPropertyName("value")]
+    public string Value { get; set; }
+    [JsonPropertyName("name")]
+    public string Name { get; set; }
+}
