+++ ReportPortal.ATM/APITests/Models/Body/WidgetOptions.cs
@@ -1,0 +1,11 @@
+﻿using System.Text.Json.Serialization;
+
+public class WidgetOptions
+{
+    [JsonPropertyName("zoom")]
+    public bool Zoom { get; set; }
+    [JsonPropertyName("timeLine")]
+    public string Timeline { get; set; }
+    [JsonPropertyName("viewMode")]
+    public string ViewMode { get; set; }
+}
